requirements.txt
openai>=1.60.0
python-dotenv>=1.0.1
tiktoken>=0.8.0
pytest>=8.0.0